# src/__generated__/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"CancelSignatureRequest": {
				"properties": {
					"signature": {
						"type": "string"
					},
					"owner_address": {
						"type": "string"
					},
					"chain_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"signature",
					"owner_address",
					"chain_id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "Hypercerts API",
		"version": "1.0.0",
		"description": "API for uploading allow lists and hypercert metadata. Visit /graphql for the GraphQL API.",
		"contact": {
			"url": "https://hypercerts.org"
		}
	},
	"paths": {
...
		"/v1/signature-requests/{safe_address}-{message_hash}/cancel": {
			"post": {
				"operationId": "CancelSignatureRequest",
				"responses": {
					"200": {
						"description": "Signature request canceled successfully",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"404": {
						"description": "Signature request not found"
					}
				},
				"tags": [
					"SignatureRequests"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "safe_address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "message_hash",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CancelSignatureRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}